[SDR-200] * SimpleRequeryRepository 구현 완료 * Cache2kEntityCache 테스트 완료 * RequeryConfiguration 테스트 완료

+++ spring-data-requery/src/main/kotlin/org/springframework/data/requery/repository/RequeryConditionExecutor.kt
@@ -31,7 +31,7 @@ interface RequeryConditionExecutor<E: Any> {
     fun findAll(conditions: Iterable<Condition<E, *>>, sort: Sort): List<E>
 
 
-    fun count(conditionElement: QueryElement<out Result<E>>): Int
+    fun count(conditionElement: QueryElement<out Result<E>>): Long
 
     fun exists(conditionElement: QueryElement<out Result<E>>): Boolean
 }
